requirements.txt
# LibCal API client
requests==2.31.0

# Environment variable management
python-dotenv==1.0.1

# Date/time handling
python-dateutil==2.8.2

# Template rendering
Jinja2==3.1.3

# JSON schema validation (optional, for config validation)
jsonschema==4.21.1
